Use jquery-ujs for delete webhook link
Previously the form for deleting a webhook was getting accidentally
triggered when hitting enter with a form field focussed. This is because
html doesn't support nesting forms.

This change uses jquery-ujs and a link for deleting the webhook. This
handles submitting as a DELETE request for us. It also has the nice side
benefit that adding a confirmation dialog is trivially easy.