Try sending webhooks 3 times before failing
This is a workaround for `PG::ConnectionBad: PQconsumeInput() SSL
connection has been closed unexpectedly` errors on Heroku which have
started occurring more frequently now that the background jobs needs to
access the database.

Ideally this wouldn't be needed because the database would be reliable,
but this seems like the simplest fix in the short term.