Make sure @countries is always defined for webhook routes
This is needed for viewing, creating and updating webhooks so that the
list of legislatures can be generated.

Previously it wasn't set on the `patch` route, which meant that when
updating a webhook failed the application would return a 500 error
because the `@countries` variable isn't set in that route.

The fix uses a `before` filter to set the instance variable correctly
for all `/webhook` routes.