Use https:// instead of git:// in Gemfile
This overrides the `git_source` definition for :github to force it to
use https://.

By default the `:github` shortcut in Bundler uses the git:// protocol,
which is insecure. This leaves us vulnerable to a man-in-the-middle
attack which could change the code as we check it out and potentially
replace it with something malicious.

See http://bundler.io/git.html#security for more details.